requirements.txt
mcp-server
pathspec